+++ src/helpers/ExpenseHelper/savedExpense.helper.ts
@@ -6,8 +6,6 @@ import { NotificationHelper } from '../BaseHelper/notification.helper';
 import { Logger } from '../BaseHelper/log.helper';
 // import { firefox } from 'playwright';
 
-let pocEmail;
-let finopsEmail;
 export class SavedExpenseCreation extends BaseHelper {
     public tabHelper: TabHelper;
     public notification: NotificationHelper;
@@ -22,7 +20,7 @@ export class SavedExpenseCreation extends BaseHelper {
     public async savedExpensePage() {
         this.locate(SavedExpenseCreation.SAVED_EXPENSE_DOM_SELECTOR);
     }
-    public async checkPartyStatus() {
+    public async checkPartyStatus(): Promise<string> {
         await this._page.waitForTimeout(1 * 5 * 1000);
         await this._page.reload();
         return await this._page
@@ -30,19 +28,12 @@ export class SavedExpenseCreation extends BaseHelper {
             .textContent();
     }
 
-    // public async toastMessage() {
-    //     return await this._page.locator("//div[@role='status']").textContent();
-    // }
-
-    // public async clickTab(buttonName: string) {
-    //     await this.click({ role: 'tab', name: buttonName });
-    // }
-    public async clickLink(linkName: string) {
+    public async clickLink(linkName: string): Promise<void> {
         await this._page.locator('a').filter({ hasText: linkName }).click();
         await this._page.waitForTimeout(1000);
     }
 
-    public async expenseStatusSuccess(statusName: string) {
+    public async expenseStatusSuccess(statusName: string): Promise<boolean> {
         const status = this._page
             .locator(`#expense-status-${statusName}`)
             .locator(`div.bg-success`);
@@ -51,7 +42,7 @@ export class SavedExpenseCreation extends BaseHelper {
             return await status.isVisible();
         }
     }
-    public async checkExpenseTo() {
+    public async checkExpenseTo(): Promise<string> {
         const toLocator = this._page.locator(
             '(//div[@aria-label="bill-to-card"]//div)[1]'
         );
@@ -61,7 +52,7 @@ export class SavedExpenseCreation extends BaseHelper {
         return toBusiness;
     }
 
-    public async checkExpenseFrom() {
+    public async checkExpenseFrom(): Promise<string> {
         const toLocator = this._page.locator(
             '(//div[@aria-label="bill-from-card"]//div)[1]'
         );
@@ -72,14 +63,14 @@ export class SavedExpenseCreation extends BaseHelper {
         return toBusiness;
     }
 
-    public async clickReject() {
+    public async clickReject(): Promise<void> {
         await this._page.getByRole('button', { name: 'Reject' }).click();
         await this.fillText('Rejected', { placeholder: 'Write a comment...' });
         await this._page.getByRole('button', { name: 'Reject' }).click();
         await this._page.waitForTimeout(1000);
     }
 
-    public async clickApprove(data: ExpenseDetailInputs[] = []) {
+    public async clickApprove(data: ExpenseDetailInputs[] = []): Promise<void> {
         const partyStatus = await this._page
             .locator("//div[@id='party-status']/div[1]/div[1]")
             .textContent();
@@ -110,234 +101,8 @@ export class SavedExpenseCreation extends BaseHelper {
     }
 }
 
-export class ApprovalWorkflowsTab extends BaseHelper {
-    private APPROVAL_WORKFLOWS_DOM_SELECTOR =
-        "//div[@aria-label='Verification Approvals']";
-
-    public async getExpData() {
-        const helper = this.locate(this.APPROVAL_WORKFLOWS_DOM_SELECTOR);
-        return helper._page
-            .locator("//span[@class='text-base text-base-secondary']")
-            .textContent();
-    }
-    public async checkLevel() {
-        const helper = this.locate(this.APPROVAL_WORKFLOWS_DOM_SELECTOR);
-        const verificationContainer = helper.locateByText('Pending Approval');
-
-        if ((await verificationContainer.count()) >= 2) {
-            return await helper._page
-                .locator("(//div[@class='col-flex']) [1]")
-                .locator('span.level-number')
-                .textContent();
-        } else {
-            return await helper._page
-                .locator("(//div[@class='col-flex'])")
-                .locator('span.level-number')
-                .textContent();
-        }
-    }
-
-    public async checkUser() {
-        const helper = this.locate(this.APPROVAL_WORKFLOWS_DOM_SELECTOR);
-        await helper._page.locator('p.approval-user-name').textContent();
-    }
-
-    public async checkEmail() {
-        const helper = this.locate(this.APPROVAL_WORKFLOWS_DOM_SELECTOR);
-        pocEmail = await helper._page
-            .locator('span.approval-user-email')
-            .textContent();
-        return pocEmail;
-    }
-    public async checkApprovalStatus(levelName) {
-        const helper = this.locate(this.APPROVAL_WORKFLOWS_DOM_SELECTOR);
-        return await helper._page
-            .getByLabel(levelName)
-            .locator('div.approval-status')
-            .first()
-            .textContent();
-    }
-    public async checkByFinOpsAdmin(labelName: string) {
-        const levelOne = this._page.locator("(//div[@class='col-flex'])[1]");
-        const levelOneVisibility = await levelOne
-            .locator(
-                "(//div[contains(@class,'row-flex justify-between')]/following-sibling::div)[1]"
-            )
-            .isVisible();
-        if (levelOneVisibility === false) {
-            await this._page
-                .getByLabel(labelName)
-                .locator('div.workflow-level-header div.icon-container')
-                .filter({ hasText: /^arrow_back_ios_new$/ })
-                .click();
-            return this._page.locator('div.approval-status').textContent();
-        } else {
-            return this._page
-                .locator('div.approval-status')
-                .first()
-                .textContent();
-        }
-    }
-
-    public async checkApprovalByFinOps() {
-        const verificationApproval = this._page.getByLabel(
-            'Verification Approvals'
-        );
-        await expect(
-            verificationApproval.locator('.approval-status'),
-            chalk.red('Approval status contains')
-        ).toContainText('Approved');
-    }
-
-    public async checkManagerApproval() {
-        const verificationApproval = this._page.getByLabel(
-            'Verification Approvals'
-        );
-        await expect(
-            verificationApproval.locator('.approval-status').nth(1)
-        ).toContainText('Approved');
-    }
-
-    public async nextVerificationFlows() {
-        await this._page.reload();
-        await this._page.reload();
-        const helper = this.locate(this.APPROVAL_WORKFLOWS_DOM_SELECTOR);
-        const verificationContainer = helper.locateByText('Pending Approval');
-        Logger.info(
-            'verificationContainer: ',
-            await verificationContainer.count()
-        );
-        for (let i; i < (await verificationContainer.count()); i++) {
-            return await verificationContainer.nth(i).textContent();
-        }
-    }
-
-    public async nextApprovalFlows(flowsName) {
-        await this._page.reload();
-
-        await this._page.reload();
-        await this._page.waitForTimeout(1000);
-        const flowsContainer = this._page.locator(
-            `//div[@aria-label='${flowsName}']`
-        );
-
-        const flowsStatus = flowsContainer
-            .locator('(//div[@class="col-flex"])')
-            .filter({ hasText: 'Pending Approval' });
-        if (flowsStatus) {
-            const email = await flowsStatus
-                .locator("(//span[@class='text-xs approval-user-email'])[2]")
-                .textContent();
-            Logger.info(email);
-            return email;
-        }
-    }
-
-    public async nextPendingFlows(flowsName) {
-        await this._page.reload();
-        await this._page.reload();
-        // await this._page.waitForLoadState('load');
-        await this._page.waitForTimeout(2000);
-        const flowsContainer = this._page
-            .locator(`//div[@aria-label='${flowsName}']`)
-            .locator('div.approval-status')
-            .nth(2);
-        if (await flowsContainer.isVisible())
-            return await flowsContainer.textContent();
-    }
-
-    public async checkPendingFlows() {
-        await this._page.reload();
-        await this._page.reload();
-        // await this._page.waitForLoadState('load');
-        await this._page.waitForTimeout(2000);
-        const verificationApproval = this._page.getByLabel('FinOps Approvals');
-        await expect(
-            verificationApproval.locator('.approval-status')
-        ).toContainText('Pending Approval');
-    }
-
-    public async clickApproveWithoutComment() {
-        await this.click({
-            selector:
-                "//div[contains(@class,'w-full border-r')]/following-sibling::button[1]",
-        });
-        await this.click({ role: 'menuitem', name: 'Approve without comment' });
-        await this._page.waitForTimeout(1000);
-    }
-}
-
-export class FinOpsVerificationHelper extends BaseHelper {
-    private FINOPS_VERIFICATION_DOM_SELECTOR =
-        '//div[@aria-label="FinOps Approvals"]';
-    helper = this.locate(this.FINOPS_VERIFICATION_DOM_SELECTOR);
-    public async getLevelName() {
-        return await this.helper._page
-            .locator('span.level-name')
-            .first()
-            .textContent();
-    }
-
-    public async getLevelStatus() {
-        return await this.helper._page
-            .locator('div.approval-status')
-            .first()
-            .textContent();
-    }
-
-    public async getFinopsEmail() {
-        finopsEmail = await this.helper._page
-            .locator('span.approval-user-email')
-            .first()
-            .textContent();
-        return finopsEmail;
-    }
-
-    public async getEmailName() {
-        return await this.helper._page
-            .locator('p.approval-user-name')
-            .first()
-            .textContent();
-    }
-}
-
-export class PaymentVerificationHelper extends BaseHelper {
-    private PAYMENT_VERIFICATION_DOM_SELECTOR =
-        '//div[@aria-label="Payment Approvals"]';
-    helper = this.locate(this.PAYMENT_VERIFICATION_DOM_SELECTOR);
-    public async getLevelName() {
-        return await this.helper._page
-            .locator('span.level-name')
-            .first()
-            .textContent();
-    }
-
-    public async getLevelStatus() {
-        return await this.helper._page
-            .locator('div.approval-status')
-            .first()
-            .textContent();
-    }
-
-    public async getPaymentEmail() {
-        finopsEmail = await this.helper._page
-            .getByLabel('Payment Approvals')
-            .locator('span.approval-user-email')
-            .first()
-            .textContent();
-        return finopsEmail;
-    }
-
-    public async getEmailName() {
-        return await this.helper._page
-            .locator('p.approval-user-name')
-            .first()
-            .textContent();
-    }
-}
-
 export class ApprovalToggleHelper extends BaseHelper {
-    public async gotoExpenseApproval() {
+    public async gotoExpenseApproval(): Promise<void> {
         // await this._page.goto(TEST_URL + '/e/f/expense-approval');
         const isExpanded = await this._page
             .locator('.hamburger_button.hamburger_button--active')
@@ -355,14 +120,14 @@ export class ApprovalToggleHelper extends BaseHelper {
             .click();
     }
 
-    public async gotoTab(tab: string) {
+    public async gotoTab(tab: string): Promise<void> {
         await this.click({
             role: 'tab',
             text: tab,
         });
     }
 
-    public async toggleOption(option: string, status: string) {
+    public async toggleOption(option: string, status: string): Promise<string> {
         const toggleRow = this._page.locator('div.table-row.body-row').filter({
             hasText: option,
         });
@@ -377,7 +142,7 @@ export class ApprovalToggleHelper extends BaseHelper {
         await button.click();
     }
 
-    public async allInactive() {
+    public async allInactive(): Promise<void> {
         await this.toggleInactive();
         await this.gotoTab('Finops');
         await this.toggleInactive();
@@ -385,7 +150,7 @@ export class ApprovalToggleHelper extends BaseHelper {
         await this.toggleInactive();
     }
 
-    public async toggleInactive() {
+    public async toggleInactive(): Promise<void> {
         await this._page.waitForSelector('div.table-row.body-row');
         const toggleRow = await this._page
             .locator('div.table-row.body-row')
